+++ R/is_sensitive.R
@@ -25,7 +25,7 @@ is_sensitive <- function( x
   r_risk <- disclosure_risk(x, risk_type = risk_type)
 
   r <- x$value
-  sensitive <- r_risk > max_risk | r$count < (x$scale*min_count)
+  sensitive <- r_risk > max_risk | r$count < (r$scale*min_count)
   names(sensitive) <- "sensitive"
   #raster::colortable(sensitive) <- c("white", "red")
   sensitive