+++ R/protect_coarsen.R
@@ -13,11 +13,11 @@ aggregate.sdc_raster <- function(x, fact = 2, ...){
     value$max <- raster::aggregate(x$value$max, fact = fact, fun = "max")
     value$max2 <- raster::aggregate(x$value$max2, fact = fact, fun = "max2")
   }
-
+  value$scale <- raster::aggregate(x$value$scale, fact = fact, fun = "mean")
   value <- value[[names(x$value)]]
 
   x_agg$value <- value
-  x_agg$scale <- 1 #TODO fix scale
+  #x_agg$scale <- 1 #TODO fix scale
   x_agg
 }